Rename camel cased file
