+++ lib/activity/index.js
@@ -9,7 +9,7 @@ const {
   pullRequestEvent,
   onStatus,
   storePRMapping,
-} = require('./pullRequests');
+} = require('./pull-requests');
 
 const { deploymentStatus } = require('./deployments');
 const { push } = require('./push');
